+++ apps/omg_eth/lib/omg_eth/blockchain/bit_helper.ex
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule Eth.Blockchain.BitHelper do
+defmodule OMG.Eth.Blockchain.BitHelper do
   @moduledoc """
   Helpers for common operations on the blockchain.
   Extracted from: https://github.com/exthereum/blockchain
