+++ apps/omg_eth/lib/omg_eth/blockchain/transaction/private_key.ex
@@ -1,0 +1,44 @@
+# Copyright 2019 OmiseGO Pte Ltd
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+defmodule OMG.Eth.Blockchain.Transaction.PrivateKey do
+  @moduledoc """
+  Extracts private key from environment
+  """
+  require Integer
+
+  def get() do
+    private_key = System.get_env("PRIVATE_KEY")
+    maybe_hex(private_key)
+  end
+
+  @spec maybe_hex(String.t() | nil) :: binary() | nil
+  defp maybe_hex(hex_data, type \\ :raw)
+  defp maybe_hex(nil, _), do: nil
+  defp maybe_hex(hex_data, :raw), do: load_raw_hex(hex_data)
+  defp maybe_hex(hex_data, :integer), do: load_hex(hex_data)
+
+  @spec load_raw_hex(String.t()) :: binary()
+  defp load_raw_hex("0x" <> hex_data), do: load_raw_hex(hex_data)
+
+  defp load_raw_hex(hex_data) when Integer.is_odd(byte_size(hex_data)),
+    do: load_raw_hex("0" <> hex_data)
+
+  defp load_raw_hex(hex_data) do
+    Base.decode16!(hex_data, case: :mixed)
+  end
+
+  @spec load_hex(String.t()) :: non_neg_integer()
+  defp load_hex(hex_data), do: hex_data |> load_raw_hex |> :binary.decode_unsigned()
+end
